requirements.txt
yfpy
pybaseball>=2.0.0
nba_api>=1.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
